included error and sucess messages and refactored edit, create and delete screen to use just one JSP

+++ src/main/java/br/com/giselepm/healthish/controller/DoctorController.java
@@ -11,9 +11,7 @@ import org.springframework.ui.ModelMap;
 import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
-import org.springframework.web.bind.annotation.RequestParam;
 
-import javax.servlet.http.HttpServletRequest;
 import java.util.List;
 
 @Controller
@@ -38,8 +36,9 @@ public class DoctorController {
 
     @RequestMapping(method = RequestMethod.GET, value = "/create")
     @Transactional(readOnly = true)
-    public String create() {
-        return "doctor/create";
+    public String create(ModelMap model) {
+        model.addAttribute("action", "Create");
+        return "doctor/show";
     }
 
     @RequestMapping(method = RequestMethod.POST, value = "/saveCreate")
@@ -49,18 +48,25 @@ public class DoctorController {
         return list(model);
     }
 
-    @RequestMapping(method = RequestMethod.POST, value = "/delete")
-    @Transactional
-    public String delete(Doctor doctor, ModelMap model) { //como eu passo o doctor inteiro na jsp?
-        crudService.delete(doctor);
-        return list(model);
-    }
-
     @RequestMapping(method = RequestMethod.GET, value = "/delete")
     @Transactional
     public String delete(Long id, ModelMap model) {
         Doctor doctor = crudService.get(Doctor.class, id);
-        crudService.delete(doctor);
+        model.addAttribute("doctor", doctor);
+        model.addAttribute("action", "Delete");
+        return "doctor/show";
+    }
+
+    @RequestMapping(method = RequestMethod.POST, value = "/saveDelete")
+    @Transactional
+    public String saveDelete(Long id, ModelMap model) {
+        Doctor doctor = crudService.get(Doctor.class, id);
+        if (doctor != null) {
+            crudService.delete(doctor);
+            model.addAttribute("doctorSuccessMsg", "The doctor was successfully deleted.");
+        } else {
+            model.addAttribute("doctorErrorMsg", "The doctor you're trying to delete doesn't exist any more.");
+        }
         return list(model);
     }
 
@@ -69,16 +75,20 @@ public class DoctorController {
     public String edit(Long id, ModelMap model) {
         Doctor doctor = crudService.get(Doctor.class, id);
         model.addAttribute("doctor", doctor);
-
-        return "doctor/edit";
+        model.addAttribute("action", "Edit");
+        return "doctor/show";
     }
 
     @RequestMapping(method = RequestMethod.POST, value = "/saveEdit")
     @Transactional
     public String saveEdit(@ModelAttribute Doctor doctor, ModelMap model) {
-        crudService.update(doctor);
+        if (crudService.get(Doctor.class, doctor.getId()) != null) {
+            crudService.update(doctor);
+            model.addAttribute("doctorSuccessMsg", "The doctor was successfully updated.");
+        } else {
+            model.addAttribute("doctorErrorMsg", "The doctor you're trying to edit doesn't exist any more.");
+        }
+
         return list(model);
     }
-
-
 }
